MYR-871: changed searching users from people
